split mellom åpne behandlinger og ferdige behandlinger.
