+++ database/src/main/kotlin/no/nav/su/se/bakover/database/DatabaseBuilder.kt
@@ -26,6 +26,7 @@ import no.nav.su.se.bakover.database.nøkkeltall.NøkkeltallPostgresRepo
 import no.nav.su.se.bakover.database.person.PersonPostgresRepo
 import no.nav.su.se.bakover.database.regulering.ReguleringPostgresRepo
 import no.nav.su.se.bakover.database.revurdering.RevurderingPostgresRepo
+import no.nav.su.se.bakover.database.sak.FerdigeBehandlingerRepo
 import no.nav.su.se.bakover.database.sak.SakPostgresRepo
 import no.nav.su.se.bakover.database.søknad.SøknadPostgresRepo
 import no.nav.su.se.bakover.database.søknadsbehandling.SøknadsbehandlingPostgresRepo
@@ -195,6 +196,7 @@ object DatabaseBuilder {
                 vedtakPostgresRepo = vedtakRepo,
                 dbMetrics = dbMetrics,
                 klageRepo = klageRepo,
+                ferdigeBehandlingerRepo = FerdigeBehandlingerRepo(),
             ),
             person = PersonPostgresRepo(
                 dataSource = dataSource,
