+++ database/src/main/kotlin/no/nav/su/se/bakover/database/sak/FerdigeBehandlingerRepo.kt
@@ -1,0 +1,201 @@
+package no.nav.su.se.bakover.database.sak
+
+import kotliquery.Row
+import no.nav.su.se.bakover.database.Session
+import no.nav.su.se.bakover.database.hentListe
+import no.nav.su.se.bakover.database.klage.KlagePostgresRepo
+import no.nav.su.se.bakover.database.revurdering.RevurderingsType
+import no.nav.su.se.bakover.database.tidspunkt
+import no.nav.su.se.bakover.domain.Saksnummer
+import no.nav.su.se.bakover.domain.sak.Behandlingsoversikt
+import no.nav.su.se.bakover.domain.søknadsbehandling.BehandlingsStatus
+import java.util.UUID
+
+internal class FerdigeBehandlingerRepo {
+
+    /**
+     * Henter ut Iverksatte, avslåtte, avsluttede behandlinger.
+     */
+    fun hentFerdigeBehandlinger(session: Session): List<Behandlingsoversikt> {
+        //language=sql
+        return """
+            with sak as (
+                select id as sakId, saksnummer
+                from sak
+            ),
+                 behandlinger as (
+                     select sak.sakId,
+                            sak.saksnummer,
+                            b.id,
+                            b.opprettet,
+                            b.status            as resultat,
+                            'SØKNADSBEHANDLING' as type,
+                            b.lukket            as avsluttet
+                     from sak
+                              join behandling b on b.sakid = sak.sakId
+                     where b.status like ('IVERKSATT%')
+                        or b.lukket = true
+                 ),
+                 revurderinger as (
+                     select sak.sakId,
+                            sak.saksnummer,
+                            r.id,
+                            r.opprettet,
+                            r.revurderingstype                                         as resultat,
+                            'REVURDERING'                                              as type,
+                            case when r.avsluttet is not null then true else false end as avsluttet
+                     from sak
+                              join behandling_vedtak bv on bv.sakid = sak.sakId
+                              join revurdering r on r.vedtaksomrevurderesid = bv.vedtakid
+                     where r.revurderingstype like ('IVERKSATT%')
+                        or r.avsluttet is not null
+                 ),
+                 klage as (
+                     select sak.sakId,
+                            sak.saksnummer,
+                            k.id,
+                            k.opprettet,
+                            k.type                                                     as resultat,
+                            'KLAGE'                                                    as type,
+                            case when k.avsluttet is not null then true else false end as avsluttet
+                     from sak
+                              join klage k on sak.sakId = k.sakid
+                     where k.type like ('iverksatt%')
+                        or k.type like 'oversendt'
+                        or k.avsluttet is not null
+                 ),
+                 slåttSammen as (
+                     select *
+                     from behandlinger
+                     union
+                     select *
+                     from revurderinger
+                     union
+                     select *
+                     from klage
+                 )
+            select *
+            from slåttSammen
+        """.trimIndent().hentListe(emptyMap(), session) {
+            it.toBehandlingsoversikt()
+        }
+    }
+
+    private fun Row.toBehandlingsoversikt(): Behandlingsoversikt {
+        val behandlignstype = BehandlingsTypeDB.valueOf(string("type"))
+
+        val erAvsluttet = boolean("avsluttet")
+
+        return Behandlingsoversikt(
+            saksnummer = Saksnummer(long("saksnummer")),
+            behandlingsId = UUID.fromString(string("id")),
+            behandlingstype = behandlignstype.toBehandlingstype(),
+            status = hentStatus(behandlignstype, erAvsluttet),
+            behandlingStartet = tidspunkt("opprettet"),
+        )
+    }
+
+    private fun Row.hentStatus(
+        restansType: BehandlingsTypeDB,
+        erBehandlingAvsluttet: Boolean,
+    ): Behandlingsoversikt.Behandlingsstatus {
+        if (erBehandlingAvsluttet) {
+            return Behandlingsoversikt.Behandlingsstatus.AVSLUTTET
+        }
+        return when (restansType) {
+            BehandlingsTypeDB.SØKNADSBEHANDLING -> BehandlingsStatus.valueOf(string("resultat"))
+                .tilSakBehandlingStatusEllerResultat()
+            BehandlingsTypeDB.REVURDERING -> RevurderingsType.valueOf(string("resultat"))
+                .tilSakBehandlingStatusEllerResultat()
+            BehandlingsTypeDB.KLAGE -> KlagePostgresRepo.Tilstand.fromString(string("resultat"))
+                .tilSakBehandlingStatusEllerResultat()
+        }
+    }
+
+    private fun BehandlingsStatus.tilSakBehandlingStatusEllerResultat(): Behandlingsoversikt.Behandlingsstatus {
+        return when (this) {
+            BehandlingsStatus.OPPRETTET,
+            BehandlingsStatus.VILKÅRSVURDERT_INNVILGET,
+            BehandlingsStatus.VILKÅRSVURDERT_AVSLAG,
+            BehandlingsStatus.BEREGNET_INNVILGET,
+            BehandlingsStatus.BEREGNET_AVSLAG,
+            BehandlingsStatus.SIMULERT,
+            BehandlingsStatus.TIL_ATTESTERING_INNVILGET,
+            BehandlingsStatus.TIL_ATTESTERING_AVSLAG,
+            BehandlingsStatus.UNDERKJENT_INNVILGET,
+            BehandlingsStatus.UNDERKJENT_AVSLAG,
+            -> throw IllegalStateException("Behandlinger som ikke er iverksatt, avslått, eller avsluttet er ikke en ferdig behandling")
+
+            BehandlingsStatus.IVERKSATT_INNVILGET -> Behandlingsoversikt.Behandlingsstatus.INNVILGET
+
+            BehandlingsStatus.IVERKSATT_AVSLAG -> Behandlingsoversikt.Behandlingsstatus.AVSLAG
+        }
+    }
+
+    private fun RevurderingsType.tilSakBehandlingStatusEllerResultat(): Behandlingsoversikt.Behandlingsstatus {
+        return when (this) {
+            RevurderingsType.OPPRETTET,
+            RevurderingsType.BEREGNET_INNVILGET,
+            RevurderingsType.BEREGNET_OPPHØRT,
+            RevurderingsType.BEREGNET_INGEN_ENDRING,
+            RevurderingsType.SIMULERT_INNVILGET,
+            RevurderingsType.SIMULERT_OPPHØRT,
+            RevurderingsType.SIMULERT_STANS,
+            RevurderingsType.SIMULERT_GJENOPPTAK,
+            RevurderingsType.TIL_ATTESTERING_INNVILGET,
+            RevurderingsType.TIL_ATTESTERING_OPPHØRT,
+            RevurderingsType.TIL_ATTESTERING_INGEN_ENDRING,
+            RevurderingsType.UNDERKJENT_INNVILGET,
+            RevurderingsType.UNDERKJENT_OPPHØRT,
+            RevurderingsType.UNDERKJENT_INGEN_ENDRING,
+            -> throw IllegalStateException("Behandlinger som ikke er iverksatt, avslått, eller avsluttet er ikke en ferdig behandling")
+
+            RevurderingsType.IVERKSATT_STANS -> Behandlingsoversikt.Behandlingsstatus.AVSLAG
+
+            RevurderingsType.IVERKSATT_OPPHØRT -> Behandlingsoversikt.Behandlingsstatus.OPPHØR
+
+            RevurderingsType.IVERKSATT_INGEN_ENDRING -> Behandlingsoversikt.Behandlingsstatus.INGEN_ENDRING
+
+            RevurderingsType.IVERKSATT_GJENOPPTAK,
+            RevurderingsType.IVERKSATT_INNVILGET,
+            -> Behandlingsoversikt.Behandlingsstatus.INNVILGET
+        }
+    }
+
+    private fun KlagePostgresRepo.Tilstand.tilSakBehandlingStatusEllerResultat(): Behandlingsoversikt.Behandlingsstatus {
+        return when (this) {
+            KlagePostgresRepo.Tilstand.OPPRETTET,
+            KlagePostgresRepo.Tilstand.VILKÅRSVURDERT_PÅBEGYNT,
+            KlagePostgresRepo.Tilstand.VILKÅRSVURDERT_UTFYLT_TIL_VURDERING,
+            KlagePostgresRepo.Tilstand.VILKÅRSVURDERT_UTFYLT_AVVIST,
+            KlagePostgresRepo.Tilstand.VILKÅRSVURDERT_BEKREFTET_TIL_VURDERING,
+            KlagePostgresRepo.Tilstand.VILKÅRSVURDERT_BEKREFTET_AVVIST,
+            KlagePostgresRepo.Tilstand.VURDERT_PÅBEGYNT,
+            KlagePostgresRepo.Tilstand.VURDERT_UTFYLT,
+            KlagePostgresRepo.Tilstand.VURDERT_BEKREFTET,
+            KlagePostgresRepo.Tilstand.AVVIST,
+            KlagePostgresRepo.Tilstand.TIL_ATTESTERING_TIL_VURDERING,
+            KlagePostgresRepo.Tilstand.TIL_ATTESTERING_AVVIST,
+            -> throw IllegalStateException("Behandlinger som ikke er iverksatt, avslått, eller avsluttet er ikke en ferdig behandling")
+
+            KlagePostgresRepo.Tilstand.OVERSENDT ->
+                Behandlingsoversikt.Behandlingsstatus.INNVILGET
+            KlagePostgresRepo.Tilstand.IVERKSATT_AVVIST ->
+                Behandlingsoversikt.Behandlingsstatus.AVSLAG
+        }
+    }
+
+    private enum class BehandlingsTypeDB {
+        SØKNADSBEHANDLING,
+        REVURDERING,
+        KLAGE;
+
+        fun toBehandlingstype(): Behandlingsoversikt.Behandlingstype {
+            return when (this) {
+                SØKNADSBEHANDLING -> Behandlingsoversikt.Behandlingstype.SØKNADSBEHANDLING
+                REVURDERING -> Behandlingsoversikt.Behandlingstype.REVURDERING
+                KLAGE -> Behandlingsoversikt.Behandlingstype.KLAGE
+            }
+        }
+    }
+}
