+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/sak/SakRepo.kt
@@ -13,6 +13,6 @@ interface SakRepo {
     fun hentSak(saksnummer: Saksnummer): Sak?
     fun hentSakIdOgNummerForIdenter(personidenter: NonEmptyList<String>): SakIdOgNummer?
     fun opprettSak(sak: NySak)
-    fun hentÅpneBehandlinger(): List<SakBehandlinger.ÅpenBehandling>
-    fun hentFerdigeBehandlinger(): List<SakBehandlinger.FerdigBehandling>
+    fun hentÅpneBehandlinger(): List<Behandlingsoversikt>
+    fun hentFerdigeBehandlinger(): List<Behandlingsoversikt>
 }
